requirements.txt
alembic==1.15.2
asgiref==3.8.1
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
click==8.1.8
distlib==0.3.9
Django==4.2.21
djangorestframework==3.16.0
filelock==3.18.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pipenv==2024.4.1
platformdirs==4.3.7
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
virtualenv==20.30.0
Werkzeug==3.1.3
zipp==3.21.0
